add : #18 UseCaseのIFを変更
