+++ src/main/kotlin/action/ShowCartItems.kt
@@ -1,5 +1,7 @@
 package action
 
-interface ShowCartItems : UseCase<Unit, String> {
+import shopping.model.cart.Cart
+
+interface ShowCartItems : UseCase<Unit, Cart> {
 
 }
